+++ docker-gen-proto.sh
@@ -23,7 +23,7 @@ gen_proto_python() {
 
 #gen_proto_dotnet accounting
 #gen_proto_java ad
-#gen_proto_dotnet cartservice
+#gen_proto_dotnet cart
 gen_proto_go checkoutservice
 #gen_proto_cpp currencyservice
 #gen_proto_ruby emailservice