+++ test/tracetesting/run.bash
@@ -8,7 +8,7 @@
 set -e
 
 # Availalble services to test
-ALL_SERVICES=("ad" "cartservice" "currencyservice" "checkoutservice" "frontend" "emailservice" "paymentservice" "productcatalogservice" "recommendationservice" "shippingservice")
+ALL_SERVICES=("ad" "cart" "currencyservice" "checkoutservice" "frontend" "emailservice" "paymentservice" "productcatalogservice" "recommendationservice" "shippingservice")
 
 ## Script variables
 # Will contain the list of services to test
@@ -36,8 +36,8 @@ spec:
   values:
     - key: AD_ADDR
       value: $AD_ADDR
-    - key: CART_SERVICE_ADDR
-      value: $CART_SERVICE_ADDR
+    - key: CART_ADDR
+      value: $CART_ADDR
     - key: CHECKOUT_SERVICE_ADDR
       value: $CHECKOUT_SERVICE_ADDR
     - key: CURRENCY_SERVICE_ADDR